Adding "BMPCLocalLauncher".
This static/utility class does something very simple but useful:
manages installation and execution of a Local BrowserMob Proxy.

A copy of BrowserMob Proxy (standalone) is stored within
the project resources.

When installed, it's deployed in a "dot-prefixed" directory
in the Current User home.

This also makes it easy to run tests for the other parts of this
library, as it doesn't need to depend on already-running
BrowserMob Proxies.